Move getCommandString to utils
